ai-qa-helper/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.4.1
pydantic==2.5.0
httpx==0.25.0
anyio==3.7.1
starlette==0.27.0
typing-extensions==4.8.0